fix(website): keep the directory anchor out of the streamed zone
The whole directory section sat under the Suspense boundary, so while the
partners API was in flight #partner-directory did not exist: the hero's
"Browse partners" jump did nothing and the profile back link landed at the
top of /partners. The boundary only became reachable in the previous commit,
so the regression shipped with it.

The section is now a server shell that renders the anchor, the heading and
the intro in the first paint and takes the streamed results as children.
Only the filter bar and the grid wait on the API, and the skeleton carries
aria-busy instead of an empty section.
