+++ packages/twenty-website/src/sections/partner-lead/index.ts
@@ -1,3 +1,4 @@
 export { PartnerBecomeStrip } from './PartnerBecomeStrip';
 export { PartnerDirectory } from './PartnerDirectory';
+export { PartnerDirectoryPanel } from './PartnerDirectoryPanel';
 export { PartnerServicesExplainer } from './PartnerServicesExplainer';
